refactor(act): group internal modules under src/internal with barrel
Move event-sourcing.ts, drain.ts, tracing.ts, and merge.ts under a new
src/internal/ directory and expose them through a single src/internal/
barrel. None of these were ever exported from the public src/index.ts
barrel — they're orchestrator-only implementation details — so grouping
them clarifies the public/internal boundary and leaves a single import
seam (`./internal/`) for the rest of src/ to depend on.

No behavior change. All consumers (act.ts, act-builder.ts,
projection-builder.ts, slice-builder.ts, plus three test files) updated
to import from `./internal/` instead of the previous flat-src locations.
