+++ libs/act/src/internal/index.ts
@@ -1,0 +1,16 @@
+/**
+ * @module internal
+ * @category Internal
+ *
+ * Barrel for framework-internal modules. These exports are not part of the
+ * public package API and may change between minors. Anything reachable from
+ * `@rotorsoft/act`'s top-level barrel (`src/index.ts`) is the public surface;
+ * what lives here is implementation detail used by `Act`, the builders, and
+ * the orchestrator's pipelines.
+ *
+ * @internal
+ */
+export * from "./drain.js";
+export * from "./event-sourcing.js";
+export * from "./merge.js";
+export * from "./tracing.js";
