+++ libs/act/src/internal/merge.ts
@@ -5,8 +5,8 @@
  * Shared utilities for merging partial states and projections across builders.
  */
 import { ZodObject, type ZodType } from "zod";
-import type { Projection } from "./projection-builder.js";
-import type { Schema, State } from "./types/index.js";
+import type { Projection } from "../projection-builder.js";
+import type { Schema, State } from "../types/index.js";
 
 /**
  * Unwraps wrapper types (ZodOptional, ZodNullable, ZodDefault, ZodReadonly)
